feat: finish makeKeys function, fix styles
