+++ src/index.js
@@ -1,9 +1,10 @@
 /* eslint-disable no-undef */
 import Keyboard from './Keyboard';
+import KeyElements from './KeyElements';
 
 // заголовок
 const header = document.createElement('h1');
-header.textContent = 'RSS Виртуальная клавиатура';
+header.textContent = 'RSS Virtual Keyboard';
 document.body.append(header);
 
 // инпут
@@ -11,13 +12,8 @@ const keyboardInput = document.createElement('textarea');
 keyboardInput.classList.add('keyboard-input');
 document.body.append(keyboardInput);
 
-// подпись
-const text = document.createElement('p');
-text.classList.add('text');
-text.textContent = 'Клавиатура создана в операционной системе Windows. Для переключения языка комбинация: левыe ctrl + alt';
-document.body.append(text);
-
-const keyboard = new Keyboard();
 window.addEventListener('DOMContentLoaded', () => {
+  const keyElements = new KeyElements(keyboardInput);
+  const keyboard = new Keyboard(keyElements.makeKeys(), keyboardInput);
   keyboard.init();
 });
