+++ src/styles/index.css
@@ -3,54 +3,76 @@
   flex-direction: column;
   align-items: center;
   justify-content: center;
+  background-color: #fff;
+  font-family: Arial, Helvetica, sans-serif;
+  color: #333030;
+}
+
+h1 {
+  margin: 20px;
+  font-size: 1.6rem;
 }
 
 .keyboard-input {
-  width: 700px;
+  width: 750px;
   height: 250px;
+  margin-bottom: 20px;
 }
 
 .text {
-  margin: 5px;
+  margin-top: 20px;
+  margin-bottom: 5px;
+  font-size: 0.8rem;
 }
 
 .keyboard {
-  width: 704px;
-  background: rgb(135, 13, 180);
+  padding: 10px;
+  border-radius: 10px;
+  width: 800px;
+  background: #fff;
+  flex-direction: row;
   text-align: center;
+  box-shadow: -2px 10px 57px 5px rgba(0, 0, 0, 0.47);
 }
 
 .keyboard__key {
-  height: 40px;
-  width: 40px;
+  height: 45px;
+  width: 45px;
   margin: 3px;
   border-radius: 4px;
   border: none;
-  background: rgba(255, 255, 255, 0.3);
-  color: #ffffff;
-  font-size: 1.1rem;
+  background: rgba(226, 153, 252, 0.9);
+  color: #333030;
+  font-size: 1rem;
   outline: none;
   cursor: pointer;
   display: inline-flex;
   align-items: center;
   justify-content: center;
   padding: 0;
+  vertical-align: top;
+}
+
+.keyboard__key:hover {
+  background: rgba(226, 153, 252);
 }
 
 .keyboard__key:active {
-  border-radius: 50%;
   background: rgba(74, 197, 33, 0.773);
 }
 
 .keyboard__key--active {
-  border-radius: 50%;
   background: rgba(74, 197, 33, 0.773);
 }
 
+.keyboard__key_medium {
+  width: 70px;
+}
+
 .keyboard__key_wide {
-  width: 90px;
+  width: 95px;
 }
 
 .keyboard__key_super-wide {
-  width: 175px;
+  width: 351px;
 }
